Added the command line and refined API
The command line was implemented, there is now a new class just for
parsing through commands and making sure that they are correct. PTrait
now randomly generates values based on a normal distribution between
-100 and 100, person has timed events to keep track of age and hunger.
Crust has most of it's object stuff completed, could do a little more.
Added a placeholder for relationships, not a lot of implementation for
it. No saving has been implemented yet.