Finished main window
The only things that has yet to be implemented is the relationships and
interests. I will start a new branch for interests. Otherwise this
should be it for the Main Window. I also moved a lot of files into their
own package so as to separate the code a bit.