Assorted things
Started the server class, mostly implemented but still needs to be
tested, changed some things on the main window so that everything is
updating correctly. Added a "Use" button, displays inventory as well.
Save is currently broken, but working on a fix so that we can save and
load everything.